feat: Reconstruct console and default_ports
