+++ src/platforms/browser/console.js
@@ -1,41 +1,17 @@
-Console = {}
+import * as _ from "../../deps/underscore-1.10.2-esm.js"
+import { inspect } from "../../system/_writer.js"
+import { Port } from "../../system/port.js"
 
-BiwaScheme.Port.current_error =
-BiwaScheme.Port.current_output = new BiwaScheme.Port.CustomOutput(
-  function (str) {
-    var console;
-    var text;
-    console = $("#bs-console");
-    if (console[0]) {
-    	text = _.escape(str);
-    	var span = $("<span>");
-    	span.html(text.replace(/\n/g,"<br>").replace(/ /g,"&nbsp;"));
-    	console.append(span);
-    }
-  }
-);
-
-BiwaScheme.Port.current_input = new BiwaScheme.Port.CustomInput(
-  function (callback) {
-    var form = $("<form/>");
-    form.html("<input id='webscheme-read-line' type='text'><input type='submit' value='ok'>");
-    $("#bs-console").append(form);
-    form.submit(function(){
-      var input = $("#webscheme-read-line").val();
-      form.remove();
-      callback(input);
-      return false;
-    });
-  }
-);
-
+const Console = {};
 
 Console.puts = function(str, no_newline) {
-  BiwaScheme.Port.current_output.put_string(str + (no_newline ? "" : "\n"))
+  Port.current_output.put_string(str + (no_newline ? "" : "\n"))
 };
 
 Console.p = function (/*ARGS*/){
-  BiwaScheme.Port.current_output.put_string(
-    "p> "+_.map(_.toArray(arguments), BiwaScheme.inspect).join(" ")
+  Port.current_output.put_string(
+    "p> "+_.map(_.toArray(arguments), inspect).join(" ")
   );
 };
+
+export default Console;
